# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CollabMate API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "in_progress",
              "done"
            ]
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teamId": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/gitlab/login": {
      "get": {
        "summary": "Redirect user to GitLab OAuth login",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to GitLab login page"
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      }
    },
    "/gitlab/callback": {
      "get": {
        "summary": "Handle GitLab OAuth callback and store access token",
        "tags": [
          "GitLab"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "description": "Authorization code from GitLab",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "description": "User ID passed as state",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GitLab account linked successfully"
          },
          "400": {
            "description": "Missing code or state"
          },
          "500": {
            "description": "GitLab OAuth failed"
          }
        }
      }
    },
    "/gitlab/user": {
      "get": {
        "summary": "Get authenticated GitLab user's profile info",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "GitLab user profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "GitLab API request failed"
          }
        }
      }
    },
    "/gitlab/projects": {
      "get": {
        "summary": "Get projects accessible to the authenticated GitLab user",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of GitLab projects retrieved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "GitLab API request failed"
          }
        }
      }
    },
    "/gitlab/unlink": {
      "delete": {
        "summary": "Unlink GitLab from the current user account",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "GitLab unlinked successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/gitlab/projects/names": {
      "get": {
        "summary": "Get user's GitLab project ID and name list",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of { id, name } for GitLab projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "GitLab not linked"
          }
        }
      }
    },
    "/gitlab/projects/{projectId}/issues/as-tasks": {
      "get": {
        "summary": "Import GitLab project issues as local tasks",
        "description": "Fetch issues from the specified GitLab project and convert them to local task format.\nAutomatically maps title, description, state, and assignees (if matched).\n",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "GitLab project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched and converted issues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "GitLab issues converted to tasks"
                    },
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "GitLab not linked"
          },
          "500": {
            "description": "GitLab API error or transformation failure"
          }
        }
      }
    },
    "/gitlab/projects/{projectId}/issues/ids": {
      "get": {
        "summary": "Get basic issue info from a GitLab project",
        "description": "Return simplified list of issueId, title, issueIid and state from GitLab project.\nUseful for preview, selection or matching.\n",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "GitLab project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of simplified issue objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "issueId": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string",
                        "enum": [
                          "opened",
                          "closed"
                        ]
                      },
                      "issueIid": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "GitLab not linked"
          },
          "500": {
            "description": "GitLab API error"
          }
        }
      }
    },
    "/gitlab/projects/{projectId}/issues/{issueIid}": {
      "get": {
        "summary": "Get single GitLab issue and convert to local task",
        "description": "Fetch one issue from GitLab by projectId and issueIid,\nand create a corresponding local task (without saving if already exists).\n",
        "tags": [
          "GitLab"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "GitLab project ID"
          },
          {
            "in": "path",
            "name": "issueIid",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "GitLab issue IID"
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully imported as task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or GitLab not linked"
          },
          "404": {
            "description": "GitLab issue not found"
          },
          "500": {
            "description": "Failed to import issue"
          }
        }
      }
    },
...
  },
...
}